requirements.txt
streamlit
pandas
scikit-learn
matplotlib
seaborn
nba-api
plotly
joblib
pytest
google-generativeai
